+++ src/app/u/components/home/profile/ConnectAccounts/AmazonConnectButton.tsx
@@ -2,9 +2,9 @@ import React from "react";
 import ConnectButton from "./ConnectButton";
 
 const AmazonConnectButton = () => {
-	return (
-		<ConnectButton unavailable={true}/>
-	);
+    return (
+        <ConnectButton unavailable={true} />
+    );
 };
 
 export default AmazonConnectButton;
