+++ src/app/u/components/home/profile/ProfileBillingPortalButton.tsx
@@ -11,49 +11,49 @@ import { useSession } from 'next-auth/react';
 
 
 const ProfileBillingPortalButton = () => {
-	const { data: session } = useSession();
-	const [billingUrl, setBillingUrl] = useState<string | null>(null);
-	const customerIdRef = useRef<string | null>(null);
-	const root = process.env.ROOT as string;
+    const { data: session } = useSession();
+    const [billingUrl, setBillingUrl] = useState<string | null>(null);
+    const customerIdRef = useRef<string | null>(null);
+    const root = process.env.ROOT as string;
 
-	useEffect(() => {
-		const fetchCheckoutUrl = async () => {
-			if (session?.user) {
+    useEffect(() => {
+        const fetchCheckoutUrl = async () => {
+            if (session?.user) {
                 const user = session.user as IUser;
-				customerIdRef.current = user.stripeCustomerId || null;
+                customerIdRef.current = user.stripeCustomerId || null;
 
-				if (customerIdRef.current) {
-					try {
-						const url = await createBillingPortalUrl(user.username ?? "", customerIdRef.current);
-						setBillingUrl(url);
-					} catch (error) {
-						console.error('Failed to create billing portal:', error);
-						setBillingUrl(root.concat('/u/failed-to-create-billing-portal'));
-					}
-				}
-			}
-		};
+                if (customerIdRef.current) {
+                    try {
+                        const url = await createBillingPortalUrl(user.username ?? "", customerIdRef.current);
+                        setBillingUrl(url);
+                    } catch (error) {
+                        console.error('Failed to create billing portal:', error);
+                        setBillingUrl(root.concat('/u/failed-to-create-billing-portal'));
+                    }
+                }
+            }
+        };
 
-		fetchCheckoutUrl();
-	}, [session, root]);
+        fetchCheckoutUrl();
+    }, [session, root]);
 
-	const handleBillingPortalButtonClick = () => {
-		if (billingUrl) {
-			window.open(billingUrl, '_blank');
-		}
-	};
+    const handleBillingPortalButtonClick = () => {
+        if (billingUrl) {
+            window.open(billingUrl, '_blank');
+        }
+    };
 
-	return (
-		<button
-			className="w-auto text-lightModeText flex items-center gap-2 pt-1 rounded-md transition duration-200 focus:outline-none hover:text-gray-500"
-			onClick={handleBillingPortalButtonClick}
-		>
-			<span className="flex items-center gap-2">
-				<TbReportMoney className="text-lg" />
-				<span className="text-base text-left">Billing Portal</span>
-			</span>
-		</button>
-	);
+    return (
+        <button
+            className="w-auto text-lightModeText flex items-center gap-2 pt-1 rounded-md transition duration-200 focus:outline-none hover:text-gray-500"
+            onClick={handleBillingPortalButtonClick}
+        >
+            <span className="flex items-center gap-2">
+                <TbReportMoney className="text-lg" />
+                <span className="text-base text-left">Billing Portal</span>
+            </span>
+        </button>
+    );
 }
 
 export default ProfileBillingPortalButton;
